QAGDEV-411-Add header and footer to the 404 and auth pages, fix styles, create constants for the routes v 2.0
