+++ src/features/kanban/constants/index.ts
@@ -4,4 +4,5 @@ export {
   STANDARD_QUERY_DEFAULTS,
   UI_CONSTANTS,
   STEP,
+  ROUTES,
 } from "./constants";
