+++ src/features/lecture-detail/constants/constants.ts
@@ -10,3 +10,7 @@ export const COMMENTS_DISPLAY_LIMIT = 3;
 export const MAX_HOMEWORK_LENGTH = 2000;
 
 export const INDEX_OFFSET = 1;
+
+export const ROUTES = {
+  TRAINING: "/training",
+};
