+++ src/features/lecture-detail/views/button-lessons-list/button-lessons-list.tsx
@@ -2,8 +2,8 @@ import { FC } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { Typography } from "@mui/material";
 
-import { StyledNavigateButton, StyledIcon } from "../buttons.styled";
-import { ROUTES } from "../../../constants/constants";
+import { ROUTES } from "../../constants";
+import { StyledNavigateButton, StyledIcon } from "./button-lessons-list.styled";
 
 const ButtonLessonsList: FC = () => {
   const { trainingId } = useParams();
