+++ src/features/lecture-detail/views/lecture-detail/lecture-detail.tsx
@@ -2,8 +2,8 @@ import { FC } from "react";
 import LectureHomework from "shared/components/lecture-homework";
 import BlurredHomework from "shared/components/blurred/blurred-homework/blurred-homework";
 import { Container } from "@mui/material";
-import { ButtonLessonsList } from "shared/components/buttons";
 
+import ButtonLessonsList from "../button-lessons-list";
 import { ILectureDetail } from "./lecture-detail.types";
 import LectureTitle from "../lecture-title";
 import LectureDescription from "../lecture-description";
