+++ src/features/training-lectures/constants/constants.ts
@@ -1,1 +1,5 @@
 export const INDEX_OFFSET = 1;
+
+export const ROUTES = {
+  HOME: "/",
+};
