+++ src/features/training-lectures/containers/training-lectures/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./training-lectures-container";
