+++ src/features/training-lectures/views/button-training-list/button-training-list.styled.ts
@@ -1,0 +1,12 @@
+import { styled } from "@mui/system";
+import { Button } from "@mui/material";
+import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
+
+export const StyledNavigateButton = styled(Button)(({ theme }) => ({
+  marginBottom: "25px",
+  color: theme.palette.primary.main,
+}));
+
+export const StyledIcon = styled(ChevronLeftIcon)({
+  marginRight: "10px",
+});
