+++ src/features/training-lectures/views/button-training-list/button-training-list.test.tsx
