+++ src/features/training-lectures/views/button-training-list/button-training-list.tsx
@@ -2,8 +2,11 @@ import { FC } from "react";
 import { useNavigate } from "react-router-dom";
 import { Typography } from "@mui/material";
 
-import { StyledNavigateButton, StyledIcon } from "../buttons.styled";
-import { ROUTES } from "../../../constants/constants";
+import {
+  StyledNavigateButton,
+  StyledIcon,
+} from "./button-training-list.styled";
+import { ROUTES } from "../../constants";
 
 const ButtonTrainingList: FC = () => {
   const navigate = useNavigate();
