+++ src/features/training-lectures/views/button-training-list/index.ts
