+++ src/features/training-lectures/views/training-lectures/index.ts
