+++ src/features/training-lectures/views/training-lectures/training-lectures.types.ts
