--- src/shared/components/buttons/index.ts
@@ -1,3 +1,0 @@
-export { default as ButtonLessonsList } from "./button-lessons-list";
-export { default as ButtonTrainingList } from "./button-training-list";
-export { default as LocalSelector } from "./local-selector";
