+++ src/shared/constants/constants.ts
@@ -16,8 +16,3 @@ export const DEFAULT_TIME_VALUE = 0;
 export const INPUT_DATE_FORMAT = "YYYY-MM-DDTHH:mm:ss";
 
 export const DATE_FORMAT = "dd.MM.yyyy '|' HH:mm";
-
-export const ROUTES = {
-  HOME: "/",
-  TRAINING: "/training",
-};
