+++ src/widgets/header/profile/profile.styled.ts
@@ -1,5 +1,5 @@
 import { styled } from "@mui/system";
-import { Box, ListItemText, MenuItem, MenuList, Stack } from "@mui/material";
+import { Box, MenuItem, MenuList, Stack } from "@mui/material";
 import Menu from "@mui/material/Menu";
 import { Link } from "react-router-dom";
 
@@ -23,7 +23,7 @@ export const StyledUserBox = styled(Box)({
 export const StyledLink = styled(Link)(({ theme }) => ({
   textDecoration: "none",
   color: theme.palette.black.main,
-  margin: "5px",
+  width: "100%",
 }));
 
 export const StyledMenuList = styled(MenuList)({
@@ -33,7 +33,7 @@ export const StyledMenuList = styled(MenuList)({
 
 export const StyledMenuItem = styled(MenuItem)(({ theme }) => ({
   margin: 0,
-  padding: "5px",
+  height: "40px",
 }));
 
 export const StyledStack = styled(Stack)(({ theme }) => ({
@@ -43,7 +43,3 @@ export const StyledStack = styled(Stack)(({ theme }) => ({
   margin: 0,
   padding: 0,
 }));
-
-export const StyledListItemText = styled(ListItemText)({
-  marginRight: "3.5vw",
-});
